refactore main file
